+++ core/runner/src/main/java/com/cognifide/aet/runner/processing/steps/CollectDispatcher.java
@@ -21,11 +21,12 @@ import com.cognifide.aet.communication.api.metadata.Url;
 import com.cognifide.aet.communication.api.queues.JmsConnection;
 import com.cognifide.aet.communication.api.wrappers.MetadataRunDecorator;
 import com.cognifide.aet.communication.api.wrappers.UrlRunWrapper;
+import com.cognifide.aet.communication.api.queues.QueuesConstant;
 import com.cognifide.aet.runner.RunnerConfiguration;
+import com.cognifide.aet.runner.processing.TimeoutWatch;
 import com.cognifide.aet.runner.processing.data.wrappers.RunIndexWrapper;
 import com.cognifide.aet.runner.scheduler.CollectorJobSchedulerService;
 import com.cognifide.aet.runner.scheduler.MessageWithDestination;
-import com.cognifide.aet.runner.processing.TimeoutWatch;
 import com.google.common.collect.Queues;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -105,7 +106,7 @@ public class CollectDispatcher extends StepManager {
 
   @Override
   protected String getQueueOutName() {
-    return "AET.collectorJobs";
+    return QueuesConstant.COLLECTOR.getJobsQueueName();
   }
 
   @Override